added ability to send mass reminder text

+++ model.go
@@ -2,5 +2,5 @@ package main
 
 type User struct {
 	PhoneNumber string
-	DayOfWeek   int
+	DayOfWeek   int64
 }
